Simplify layout (#489)
